t210b01: abca: add "nvidia,active-panel" property
As per the redesign of the driver, active display panel on the given
platform will be selected by property "nvidia,active-panel". Driver
first searches the active panel based on display board-id and if
it doesn't find one, it will look for this property for the
active panel on the given platform.

TDS-1544

Change-Id: I4d3f1a528a93586f5b296451fe82f630d0f3ad91
Signed-off-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-on: http://git-master/r/1328213
Reviewed-by: Shu Zhong <shuz@nvidia.com>
GVS: Gerrit_Virtual_Submit