DT: t210b01: abca: Use pinctrl framework for initialising IO pad voltages
PMC driver register the pinctrl framework to provide the
initialisation of IO pad voltages during init.

Move to use the pinctrl framework to initialise the IO pad
voltages and get rid of the legacy way of initialisation
via pad control driver which is going to be deprecated.

Change-Id: If63ef4a12bde4ebb5c16feaa5c92424306856c5c
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/1466747
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Venkat Reddy Talla <vreddytalla@nvidia.com>